fix test and refactor members
