+++ pages/api/spaces/[id]/members/index.ts
@@ -3,13 +3,11 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import nc from 'next-connect';
 
 import { prisma } from 'db';
-import { getAccessibleMemberPropertiesBySpace } from 'lib/members/getAccessibleMemberPropertiesBySpace';
+import { getSpaceMembers } from 'lib/members/getSpaceMembers';
 import type { Member, PublicMember } from 'lib/members/interfaces';
-import { getPropertiesWithValues } from 'lib/members/utils';
 import { onError, onNoMatch } from 'lib/middleware';
 import { withSessionRoute } from 'lib/session/withSession';
 import { hasAccessToSpace } from 'lib/users/hasAccessToSpace';
-import { hasNftAvatar } from 'lib/users/hasNftAvatar';
 import { deterministicRandomName } from 'lib/utilities/randomName';
 
 const handler = nc<NextApiRequest, NextApiResponse>({ onError, onNoMatch });
@@ -112,75 +110,7 @@ async function getMembers(req: NextApiRequest, res: NextApiResponse<(Member | Pu
     });
   }
 
-  const spaceRoles = await prisma.spaceRole.findMany({
-    where:
-      whereOr.length !== 0
-        ? {
-            spaceId,
-            OR: whereOr
-          }
-        : {
-            spaceId
-          },
-    include: {
-      user: {
-        include: {
-          profile: true,
-          memberPropertyValues: {
-            where: {
-              spaceId
-            }
-          }
-        }
-      },
-      spaceRoleToRole: {
-        include: {
-          role: {
-            select: {
-              id: true,
-              name: true
-            }
-          }
-        }
-      }
-    }
-  });
-
-  const visibleProperties = await getAccessibleMemberPropertiesBySpace({ requestingUserId: userId, spaceId });
-
-  const members = spaceRoles
-    .map((spaceRole): Member => {
-      const { memberPropertyValues = [], id, ...userData } = spaceRole.user;
-      const roles = spaceRole.spaceRoleToRole?.map((sr) => sr.role);
-
-      const nameProperty = visibleProperties.find((property) => property.type === 'name') ?? null;
-      const memberNameProperty = memberPropertyValues.find((prop) => prop.memberPropertyId === nameProperty?.id);
-      return {
-        id,
-        ...userData,
-        username: memberNameProperty?.value || userData.username,
-        addresses: [],
-        onboarded: spaceRole.onboarded,
-        isAdmin: spaceRole.isAdmin,
-        joinDate: spaceRole.createdAt.toISOString(),
-        hasNftAvatar: hasNftAvatar(spaceRole.user),
-        properties: getPropertiesWithValues(visibleProperties, memberPropertyValues),
-        roles
-      } as Member;
-    })
-    .filter((member) => !member.deletedAt) // filter out deleted members
-    .sort((a, b) => {
-      const first = a.username.toLowerCase();
-      const second = b.username.toLowerCase();
-
-      if (first < second) {
-        return -1;
-      } else if (second > first) {
-        return 1;
-      } else {
-        return 0;
-      }
-    }); // sort members alphabetically
+  const members = await getSpaceMembers({ spaceId, requestingUserId: userId, whereOr });
 
   return res.status(200).json(members);
 }
